requirements.txt
python -m venv .venv
source .venv/bin/activate        # or .venv\Scripts\activate on Windows
pip install django python-dotenv
django-admin startproject myproject .
git init
